rename commands to _commands
