+++ asdf/_tests/commands/test_tags.py
@@ -3,7 +3,7 @@ import io
 import pytest
 
 from asdf import AsdfFile
-from asdf.commands import list_tags
+from asdf._commands import list_tags
 
 
 @pytest.mark.parametrize("display_classes", [True, False])
